"Setup base project structure and packages for rust"
Added essential components for the start of a new Rust project. This includes a base 'mod.rs' file in 'src/solutions' for module declarations, a 'Cargo.toml' for package management, and a 'main.rs' file for the entry point of the application. Also, '.idea' has been added to gitignore. This setup is important for initiating the application building process in Rust.