Add string to array conversion function and tests
Added a new module 'convert_a_string_to_an_array' with a function to convert a given string into an array by splitting the string by spaces. This function is useful in various situations like text processing and tokenization. Unit tests were added to validate the correctness of this function with multiple sets of input, including multi-word strings and single-word strings. This new module was also included in the 'solutions' module index.