Add beginner_reduce_but_grow module and tests
Created a new function 'grow' in the new module 'beginner_reduce_but_grow'. The 'grow' function takes a list of integers and returns the product of all the integers in the list. It includes a series of tests to ensure the function works as expected. Lastly, this new module has been added to the 'solutions' module index.